feat: implementando tratamento de exceções

+++ src/main/java/dev/afinovicz/ecommerce_project/services/ProductService.java
@@ -3,11 +3,17 @@ package dev.afinovicz.ecommerce_project.services;
 import dev.afinovicz.ecommerce_project.dto.ProductDTO;
 import dev.afinovicz.ecommerce_project.entities.Product;
 import dev.afinovicz.ecommerce_project.repositories.ProductRepository;
+import dev.afinovicz.ecommerce_project.services.exceptions.DatabaseException;
+import dev.afinovicz.ecommerce_project.services.exceptions.ResourceNotFoundException;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import tools.jackson.databind.exc.InvalidDefinitionException;
 
 import java.util.Optional;
 
@@ -19,10 +25,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO productDTO = new ProductDTO(product);
-        return productDTO;
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado!"));
+        return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -41,15 +46,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+         catch(ObjectNotFoundException e) {
+             throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial!");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
